+++ src/app/shared/shared.module.ts
@@ -2,16 +2,19 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { BackgroundColorDirective } from './directives/background-color.directive';
+import { BorderDirective } from './directives/border.directive';
 
 @NgModule({
   declarations: [
-    BackgroundColorDirective
+    BackgroundColorDirective,
+    BorderDirective
   ],
   imports: [
     CommonModule
   ],
   exports: [
-    BackgroundColorDirective
+    BackgroundColorDirective,
+    BorderDirective
   ]
 })
 export class SharedModule { }